[TS-179] Fixes important API classes

+++ src/de/codingair/tradesystem/utils/Lang.java
@@ -97,7 +97,7 @@ public class Lang {
             }
 
             to.write(buf, 0, r);
-            total += (long) r;
+            total += r;
         }
     }
 }
